pyproject.toml
[project]
name = "adk-langchain-comparison"
version = "0.1.0"
description = "Comparison between Google ADK and LangChain multi-agent patterns"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "google-genai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

